@@ -1,49 +1,14 @@
 import { Construct, IConstruct, Node } from 'constructs';
 import { Chart } from './chart';
-import { removeEmpty } from './_util';
+import { sanitizeValue } from './_util';
 import { resolve } from './_tokens';
 import * as stringify from 'json-stable-stringify';
-
-/**
- * Metadata associated with this object.
- */
-export interface ApiObjectMetadata {
-  /**
-   * The unique, namespace-global, name of this object inside the Kubernetes
-   * cluster.
-   *
-   * Normally, you shouldn't specify names for objects and let the CDK generate
-   * a name for you that is application-unique. The names CDK generates are
-   * composed from the construct path components, separated by dots and a suffix
-   * that is based on a hash of the entire path, to ensure uniqueness.
-   *
-   * You can supply custom name allocation logic by overriding the
-   * `chart.generateObjectName` method.
-   *
-   * If you use an explicit name here, bear in mind that this reduces the
-   * composability of your construct because it won't be possible to include
-   * more than one instance in any app. Therefore it is highly recommended to
-   * leave this unspecified.
-   *
-   * @default - an app-unique name generated by the chart
-   */
-  readonly name?: string;
-
-  /**
-   * Arbitrary key/value metadata.
-   */
-  readonly [key: string]: any;
-}
+import { ApiObjectMetadata, ApiObjectMetadataDefinition } from './metadata';
 
 /**
  * Options for defining API objects.
  */
 export interface ApiObjectOptions {
-  /**
-   * Data associated with the resource.
-   */
-  readonly data?: any;
-
   /**
    * Object metadata.
    *
@@ -95,6 +60,11 @@ export class ApiObject extends Construct {
   public readonly chart: Chart;
 
   /**
+   * Metadata associated with this API object.
+   */
+  public readonly metadata: ApiObjectMetadataDefinition;
+
+  /**
    * Defines an API object.
    *
    * @param scope the construct scope
@@ -106,7 +76,16 @@ export class ApiObject extends Construct {
     this.chart = Chart.of(this);
     this.kind = options.kind;
     this.apiVersion = options.apiVersion;
+
     this.name = options.metadata?.name ?? this.chart.generateObjectName(this);
+
+    this.metadata = new ApiObjectMetadataDefinition({
+      name: this.name,
+      namespace: options.metadata?.namespace ?? this.chart.namespace,
+
+      // override user defined values
+      ...options.metadata,
+    });
   }
 
   /**
@@ -125,15 +104,11 @@ export class ApiObject extends Construct {
   public toJson(): any {
     const data = {
       ...this.options,
-      metadata: {
-        namespace: Chart.of(this).namespace,
-        ...this.options.metadata,
-        name: this.name,
-      }
+      metadata: this.metadata.toJson(),
     };
 
     // convert to "pure data" so, for example, when we convert to yaml these
     // references are not converted to anchors.
-    return JSON.parse(stringify(removeEmpty(resolve(this, data))));
+    return JSON.parse(stringify(sanitizeValue(resolve(this, data))));
   }
 }
